Allow to trigger on events other than `push` and secure the webhook (#39)
This makes the bot more flexible and responsive. We'll run backports only on `push` events to `main`.

Label maintenance and merge queue maintenance will be not only triggered only on `push` events but also on pull requests label events. To make sure we don't do too much unnecessary work, we only trigger these when either a `reviewed` or a `backport` label has been applied.

A new environment variable is introduced, `BACKPORTER_GITHUB_SECRET`. It'll be used to verify the requests are sent from GitHub.

- Closes #9 
- Part 1 of #38 (this change is required for that)

An octokit dependency is introduced for webhooks, we can later use the octokit SDK to query GitHub.