Replace LGTM bot (#51)
This will do the same thing the LGTM bot did more or less. We won't parse the `MAINTAINERS` file nor look for comments containing the word LGTM.

This is experimental as I couldn't really test it. Before this is merged, the old LGTM bot must be stopped. After this is merged, the context `backporter/lgtm` must be added to `go-gitea/gitea`'s required status checks in all protected branches (and `approvals/lgtm` removed).
![image](https://user-images.githubusercontent.com/20454870/233852721-5d6950ca-f520-48d4-9144-bd80a08c0c53.png)

If the `BACKPORTER_GITHUB_TOKEN` used for production doesn't contain the `repo:status` scope, it must be added before merging.
![image](https://user-images.githubusercontent.com/20454870/233853011-bbf86efc-bbfb-413c-ae14-364a26297ee3.png)

- Closes #50